Move ResourceHelper from util to server subpackage
ResourceHelper is used exclusively by server-side components
(MiniProfilerServer, ProfilingFilter, MiniProfilerResourceHandler).
Place it alongside the other server utilities it supports.
